chore: domain 패키지 재정의 command/entity/exception/repository/result/service/support
