+++ src/main/java/matchuri/backend/domain/member/command/RegisterLocalMemberCommand.java
@@ -1,4 +1,4 @@
-package matchuri.backend.domain.member.service;
+package matchuri.backend.domain.member.command;
 
 import java.util.List;
