+++ src/main/java/matchuri/backend/domain/member/command/UpdateMemberBasicInfoCommand.java
@@ -1,4 +1,4 @@
-package matchuri.backend.domain.member.service;
+package matchuri.backend.domain.member.command;
 
 public record UpdateMemberBasicInfoCommand(
         String nickname
