+++ src/main/java/matchuri/backend/domain/member/result/CreateMemberResult.java
@@ -1,11 +1,10 @@
-package matchuri.backend.domain.member.service;
+package matchuri.backend.domain.member.result;
 
 import java.time.LocalDateTime;
 
-public record RegisterLocalMemberResult(
+public record CreateMemberResult(
         Long memberId,
         String loginId,
-        String nickname,
         LocalDateTime createdAt
 ) {
 }
