+++ src/main/java/matchuri/backend/domain/member/result/MemberProfileResult.java
@@ -1,4 +1,4 @@
-package matchuri.backend.domain.member.service;
+package matchuri.backend.domain.member.result;
 
 public record MemberProfileResult(
         Long id,
