+++ src/main/java/matchuri/backend/domain/member/result/SubmitRequiredAgreementsResult.java
@@ -1,4 +1,4 @@
-package matchuri.backend.domain.member.service;
+package matchuri.backend.domain.member.result;
 
 import matchuri.backend.domain.auth.service.IssuedAccessToken;
