+++ src/main/java/matchuri/backend/domain/member/result/WithdrawMemberResult.java
@@ -1,4 +1,4 @@
-package matchuri.backend.domain.member.service;
+package matchuri.backend.domain.member.result;
 
 public record WithdrawMemberResult(
         Long id,
