+++ src/main/java/matchuri/backend/domain/member/service/MemberAgreementService.java
@@ -1,5 +1,9 @@
 package matchuri.backend.domain.member.service;
 
+import matchuri.backend.domain.member.command.SubmitRequiredAgreementsCommand;
+import matchuri.backend.domain.member.result.RequiredAgreementStatusResult;
+import matchuri.backend.domain.member.result.SubmitRequiredAgreementsResult;
+
 public interface MemberAgreementService {
 
     RequiredAgreementStatusResult getRequiredAgreementStatus();
