+++ src/main/java/matchuri/backend/domain/member/support/member/ActiveMemberReader.java
@@ -1,0 +1,37 @@
+package matchuri.backend.domain.member.support.member;
+
+import lombok.RequiredArgsConstructor;
+import matchuri.backend.domain.member.entity.Member;
+import matchuri.backend.domain.member.entity.MemberStatus;
+import matchuri.backend.domain.member.exception.MemberErrorCode;
+import matchuri.backend.domain.member.repository.MemberRepository;
+import matchuri.backend.global.exception.BusinessException;
+import matchuri.backend.global.security.AuthenticatedMember;
+import matchuri.backend.global.security.AuthenticationFacade;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class ActiveMemberReader {
+
+    private final MemberRepository memberRepository;
+    private final AuthenticationFacade authenticationFacade;
+
+    public Member getCurrentAuthenticatedActiveMember() {
+        AuthenticatedMember authenticatedMember = authenticationFacade.getCurrentMember();
+        Member member = memberRepository.findById(authenticatedMember.memberId())
+                .orElseThrow(() -> new BusinessException(
+                        MemberErrorCode.NOT_FOUND,
+                        MemberErrorCode.NOT_FOUND.format(authenticatedMember.memberId())
+                ));
+
+        if (member.getStatus() != MemberStatus.ACTIVE) {
+            throw new BusinessException(
+                    MemberErrorCode.INACTIVE_MEMBER,
+                    MemberErrorCode.INACTIVE_MEMBER.format(member.getId())
+            );
+        }
+
+        return member;
+    }
+}
