+++ src/main/java/matchuri/backend/global/security/RequiredAgreementAccessFilter.java
@@ -8,9 +8,9 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import matchuri.backend.domain.member.MemberAgreementErrorCode;
-import matchuri.backend.domain.member.service.RequiredAgreementRevisionResolver;
-import matchuri.backend.domain.member.service.RequiredAgreementVersions;
+import matchuri.backend.domain.member.exception.MemberAgreementErrorCode;
+import matchuri.backend.domain.member.support.agreement.RequiredAgreementRevisionResolver;
+import matchuri.backend.domain.member.support.agreement.RequiredAgreementVersions;
 import matchuri.backend.global.config.MatchuriProperties;
 import org.springframework.http.HttpMethod;
 import org.springframework.security.access.AccessDeniedException;
